https://gaiaplatform.atlassian.net/browse/GAIAPLAT-1185 : Added ability to specify new package to install.

+++ test/reset_database.sh
@@ -113,7 +113,7 @@ stop_database_service() {
         echo "Service $service_name is currently $service_state."
     fi
 
-    if ! [ "$service_state" == "inactive" ] ; then
+    if [ "$service_state" != "inactive" ] && [ "$service_state" != "failed" ]; then
         echo "Setting service $service_name to 'inactive'."
         if ! sudo systemctl stop gaia ; then
             echo "Service $service_name cannot be stopped."
@@ -163,6 +163,40 @@ remove_data_store() {
     fi
 }
 
+# If an install file was specified, make sure it is present and looks like a debian install file.
+verify_install_file_if_present() {
+    INSTALL_FILE=${PARAMS[0]}
+    if [ -n "$INSTALL_FILE" ] ; then
+
+        # Make sure it exists.
+        if [ ! -f "$INSTALL_FILE" ]; then
+            echo "Install file '$(realpath "$INSTALL_FILE")' does not exist."
+            complete_process 1
+        fi
+
+        if [[ ! "$INSTALL_FILE" == *.deb ]] ; then
+            echo "Install file '$(realpath "$INSTALL_FILE")' does not end with the required suffix '.deb'."
+            complete_process 1
+        fi
+
+        echo "Installation of new packages requires resetting of the database.  Option selected."
+        RESET_DATABASE=1
+    fi
+}
+
+# Remove the old package and install the new package.
+install_new_package() {
+    if ! sudo apt --assume-yes remove gaia ; then
+        echo "Removal of old Gaia package did not complete.  Gaia may be in an undefined state."
+        complete_process 1
+    fi
+
+    if ! sudo apt --assume-yes install "$INSTALL_FILE" ; then
+        echo "Installation of new Gaia package did not complete.  Gaia may be in an undefined state."
+        complete_process 1
+    fi
+}
+
 # Set up any global script variables.
 # shellcheck disable=SC2164
 #SCRIPTPATH="$( cd -- "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
@@ -177,6 +211,9 @@ service_name=gaia
 # Parse any command line values.
 parse_command_line "$@"
 
+verify_install_file_if_present
+
+
 # Clean entrance into the script.
 start_process
 
@@ -185,8 +222,10 @@ stop_database_service
 if [ $RESET_DATABASE -ne 0 ] ; then
     remove_data_store
 fi
-#sudo apt remove gaia
-#sudo apt install ./2021Aug05-gaia-0.1.0_amd64.deb
+
+if [ -n "$INSTALL_FILE" ] ; then
+    install_new_package
+fi
 
 start_database_service
